Create/Manipulate Hash Keys

as_key - Create your own hash keys from a data frame for use in key arguments such as polarity_dt in the sentiment function.

update_key - Add/remove terms to a current key.

update_polarity_table - Wrapper for update_key specifically for updating polarity tables.

update_valence_shifter_table - Wrapper for update_key specifically for updating valence shifter tables.

is_key - Logical check if an object is a key.

Usage

as_key(x, comparison = lexicon::hash_valence_shifters, sentiment = TRUE, ...)

update_key(
  key,
  drop = NULL,
  x = NULL,
  comparison = lexicon::hash_valence_shifters,
  sentiment = FALSE,
  ...
)

update_polarity_table(
  key,
  drop = NULL,
  x = NULL,
  comparison = lexicon::hash_valence_shifters,
  sentiment = FALSE,
  ...
)

update_valence_shifter_table(
  key,
  drop = NULL,
  x = NULL,
  comparison = lexicon::hash_sentiment_jockers_rinker,
  sentiment = FALSE,
  ...
)

is_key(key, sentiment = TRUE)

Arguments

x

A data.frame with the first column containing polarized words and the second containing polarity values.

comparison

A data.frame to compare to x. If elements in x's column 1 matches comparison's column 1 the accompanying row will be removed from x. This is useful to ensure polarity_dt words are not also found in valence_shifters_dt in sentiment. Use comparison = NULL to skip this comparison.

sentiment

logical. If TRUE checking expects column 2 of the input keys/data.frame are expected to be numeric.

key

A sentimentr hash key.

drop

A vector of terms to drop.

...

ignored.

Value

Returns a data.table object that can be used as a hash key.

Details

For updating keys via update_key note that a polarity_dt and valence_shifters_dt are the primary dictionary keys used in the sentimentr package. The polarity_dt takes a 2 column data.frame (named x and y) with the first column being character and containing the words and the second column being numeric values that are positive or negative. valence_shifters_dt takes a 2 column data.frame (named x and y) with the first column being character and containing the words and the second column being integer corresponding to: (1) negators, (2) amplifiers, (3) de-amplifiers, and (4) dversative conjunctions (i.e., 'but', 'however', and 'although'). Also, note that if you are updating a valence_shifters_dt you need an appropriate comparison; most likely, comparison = sentimentr::polarity_dt.

Examples

key <- data.frame(
    words = sample(letters),
    polarity = rnorm(26),
    stringsAsFactors = FALSE
)

(mykey <- as_key(key))
#>     x            y
#>  1: a -0.778045742
#>  2: b  0.565243652
#>  3: c -1.811767932
#>  4: d  2.319703217
#>  5: e -1.058141604
#>  6: f -0.450962614
#>  7: g  0.155787329
#>  8: h  0.317620484
#>  9: i -0.398177623
#> 10: j  0.709554609
#> 11: k  0.976963810
#> 12: l -0.944796093
#> 13: m  0.070513448
#> 14: n  0.671526601
#> 15: o -0.704242110
#> 16: p -0.637526722
#> 17: q -2.003736924
#> 18: r -1.616070441
#> 19: s -1.445045642
#> 20: t -0.865754738
#> 21: u -1.421267299
#> 22: v  0.008290401
#> 23: w -1.764894382
#> 24: x -1.096107126
#> 25: y  0.207028624
#> 26: z -2.469637641
#>     x            y

## Looking up values
mykey[c("a", "k")][[2]]
#> [1] -0.7780457  0.9769638

## Drop terms from key
update_key(mykey, drop = c("f", "h"))
#>     x            y
#>  1: a -0.778045742
#>  2: b  0.565243652
#>  3: c -1.811767932
#>  4: d  2.319703217
#>  5: e -1.058141604
#>  6: g  0.155787329
#>  7: i -0.398177623
#>  8: j  0.709554609
#>  9: k  0.976963810
#> 10: l -0.944796093
#> 11: m  0.070513448
#> 12: n  0.671526601
#> 13: o -0.704242110
#> 14: p -0.637526722
#> 15: q -2.003736924
#> 16: r -1.616070441
#> 17: s -1.445045642
#> 18: t -0.865754738
#> 19: u -1.421267299
#> 20: v  0.008290401
#> 21: w -1.764894382
#> 22: x -1.096107126
#> 23: y  0.207028624
#> 24: z -2.469637641
#>     x            y

## Add terms to key
update_key(mykey, x = data.frame(x = c("dog", "cat"), y = c(1, -1)))
#>       x            y
#>  1:   a -0.778045742
#>  2:   b  0.565243652
#>  3:   c -1.811767932
#>  4: cat -1.000000000
#>  5:   d  2.319703217
#>  6: dog  1.000000000
#>  7:   e -1.058141604
#>  8:   f -0.450962614
#>  9:   g  0.155787329
#> 10:   h  0.317620484
#> 11:   i -0.398177623
#> 12:   j  0.709554609
#> 13:   k  0.976963810
#> 14:   l -0.944796093
#> 15:   m  0.070513448
#> 16:   n  0.671526601
#> 17:   o -0.704242110
#> 18:   p -0.637526722
#> 19:   q -2.003736924
#> 20:   r -1.616070441
#> 21:   s -1.445045642
#> 22:   t -0.865754738
#> 23:   u -1.421267299
#> 24:   v  0.008290401
#> 25:   w -1.764894382
#> 26:   x -1.096107126
#> 27:   y  0.207028624
#> 28:   z -2.469637641
#>       x            y

## Add terms & drop to/from a key
update_key(mykey, drop = c("f", "h"), x = data.frame(x = c("dog", "cat"), y = c(1, -1)))
#>       x            y
#>  1:   a -0.778045742
#>  2:   b  0.565243652
#>  3:   c -1.811767932
#>  4: cat -1.000000000
#>  5:   d  2.319703217
#>  6: dog  1.000000000
#>  7:   e -1.058141604
#>  8:   g  0.155787329
#>  9:   i -0.398177623
#> 10:   j  0.709554609
#> 11:   k  0.976963810
#> 12:   l -0.944796093
#> 13:   m  0.070513448
#> 14:   n  0.671526601
#> 15:   o -0.704242110
#> 16:   p -0.637526722
#> 17:   q -2.003736924
#> 18:   r -1.616070441
#> 19:   s -1.445045642
#> 20:   t -0.865754738
#> 21:   u -1.421267299
#> 22:   v  0.008290401
#> 23:   w -1.764894382
#> 24:   x -1.096107126
#> 25:   y  0.207028624
#> 26:   z -2.469637641
#>       x            y

## Explicity key type (wrapper for `update_key` for sentiment table.
## See `update_valence_shifter_table` a corresponding valence shifter updater.
library(lexicon)
#> 
#> Attaching package: ‘lexicon’
#> The following object is masked from ‘package:sentimentr’:
#> 
#>     available_data
updated_hash_sentiment <- sentimentr:::update_polarity_table(lexicon::hash_sentiment_huliu,
    x = data.frame(
        words = c('frickin', 'hairy'),
        polarity = c(-1, -1),
        stringsAsFactors = FALSE
    )
)

## Checking if you have a key
is_key(mykey)
#> [1] TRUE
is_key(key)
#> [1] FALSE
is_key(mtcars)
#> [1] FALSE
is_key(update_key(mykey, drop = c("f", "h")))
#> [1] TRUE

## Using syuzhet's sentiment lexicons
if (FALSE) {
library(syuzhet)
(bing_key <- as_key(syuzhet:::bing))
as_key(syuzhet:::afinn)
as_key(syuzhet:::syuzhet_dict)

sam <- gsub("Sam-I-am", "Sam I am", sam_i_am)
sentiment(sam, , polarity_dt = bing_key)

## The nrc dictionary in syuzhet requires a bit of data wrangling before it 
## is in the correct shape to convert to a key.  

library(syuzhet)
library(tidyverse)

nrc_key <- syuzhet:::nrc %>% 
    dplyr::filter(
        sentiment %in% c('positive', 'negative'),
        lang == 'english'
    ) %>%
    dplyr::select(-lang) %>% 
    mutate(value = ifelse(sentiment == 'negative', value * -1, value)) %>%
    dplyr::group_by(word) %>%
    dplyr::summarize(y = mean(value)) %>%
    sentimentr::as_key()
    
sentiment(sam, polarity_dt = nrc_key)

## The lexicon package contains a preformatted nrc sentiment hash table that 
## can be used instead.
sentiment(sam, polarity_dt = lexicon::hash_sentiment_nrc)
}

## Using 2 vectors of words
if (FALSE) {
install.packages("tm.lexicon.GeneralInquirer", repos="http://datacube.wu.ac.at", type="source")
require("tm.lexicon.GeneralInquirer")

positive <- terms_in_General_Inquirer_categories("Positiv")
negative <- terms_in_General_Inquirer_categories("Negativ")

geninq <- data.frame(
    x = c(positive, negative),
    y = c(rep(1, length(positive)), rep(-1, length(negative))),
    stringsAsFactors = FALSE
) %>%
    as_key()

geninq_pol <- with(presidential_debates_2012,
    sentiment_by(dialogue,
    person,
    polarity_dt = geninq
))

geninq_pol %>% plot()
}