Combine sentimentr's Sentiment Data Sets

Combine trusted sentiment data sets from sentimentr.

Usage

combine_data(
  data = c("course_evaluations", "hotel_reviews", "kaggle_movie_reviews",
    "kotzias_reviews_amazon_cells", "kotzias_reviews_imdb", "kotzias_reviews_yelp",
    "nyt_articles"),
  ...
)

Arguments

data

A character vector of sentimentr data sets.

...

ignored.

Value

Returns an rbinded data.table of sentiment data with the source added as column.

Examples

combine_data()
#>                    source sentiment
#>     1: course_evaluations     -1.00
#>     2: course_evaluations     -1.00
#>     3: course_evaluations      1.00
#>     4: course_evaluations      1.00
#>     5: course_evaluations     -1.00
#>    ---                             
#> 20975:       nyt_articles      0.45
#> 20976:       nyt_articles      0.60
#> 20977:       nyt_articles      0.70
#> 20978:       nyt_articles      0.65
#> 20979:       nyt_articles      0.15
#>                                                                                                                                                                                                             text
#>     1:                                                                                                                                   The labs in EECS 588 aren't very engaging or challenging in my opinion.
#>     2:                                                                                                                                                                                   I didn't like EECS 417.
#>     3:                                                                                                                                       My three favorite classes in EECS have been EECS 230, 498, and 583.
#>     4:                                                                                                                                Foundations of computer science looks difficult but useful and insightful.
#>     5:                                                                                I have learned more in one month of Mobile App Development for Entrepreneurs than I learned in all of Electromagnetics II.
#>    ---                                                                                                                                                                                                          
#> 20975: The report highlights familiar remedies: decentralize bureaucracies, make schools more accountable, break up large schools into smaller units and build partnerships with parents, colleges and business.
#> 20976:                                                                                                                                  It also urges greater equity in state funding of local school districts.
#> 20977:                                               It proposes more Federal funding for a range of programs like early childhood nutrition and incentives for closer ties between colleges and ghetto schools.
#> 20978:                                    ''The quality of our education will determine the strength of our democracy, the vitality of our economy and the promise of our ideals,'' the Carnegie trustees write.
#> 20979:                                                                                                                                                   They sound a needed alarm in the urban education night.
combine_data(c("kotzias_reviews_amazon_cells", "kotzias_reviews_imdb", 
    "kotzias_reviews_yelp"))
#>                             source sentiment
#>    1: kotzias_reviews_amazon_cells        -1
#>    2: kotzias_reviews_amazon_cells         1
#>    3: kotzias_reviews_amazon_cells         1
#>    4: kotzias_reviews_amazon_cells        -1
#>    5: kotzias_reviews_amazon_cells        -1
#>   ---                                       
#> 3144:         kotzias_reviews_yelp        -1
#> 3145:         kotzias_reviews_yelp        -1
#> 3146:         kotzias_reviews_yelp        -1
#> 3147:         kotzias_reviews_yelp        -1
#> 3148:         kotzias_reviews_yelp        -1
#>                                                                                                                                         text
#>    1:                                                     So there is no way for me to plug it in here in the US unless I go by a converter.
#>    2:                                                                                                            Good case, Excellent value.
#>    3:                                                                                                                 Great for the jawbone.
#>    4:                                                                        Tied to charger for conversations lasting more than 45 minutes.
#>    5:                                                                                                                       MAJOR PROBLEMS!!
#>   ---                                                                                                                                       
#> 3144:                                                                     I think food should have flavor and texture and both were lacking.
#> 3145:                                                                                                               Appetite instantly gone.
#> 3146:                                                                                     Overall I was not impressed and would not go back.
#> 3147:                                            The whole experience was underwhelming, and I think we'll just go to Ninja Sushi next time.
#> 3148: Then, as if I hadn't wasted enough of my life there, they poured salt in the wound by drawing out the time it took to bring the check.