Rescale a Numeric Vector

Rescale a numeric vector with the option to make signed (-1, 1, or 0) and retain zero as neutral.

Usage

general_rescale(
  x,
  lower = -1,
  upper = 1,
  mute = NULL,
  keep.zero = lower < 0,
  sign = FALSE,
  ...
)

Arguments

x

A numeric vector.

lower

An upper limit to rescale to.

upper

A lower limit to rescale to.

mute

A positive value greater than 1 to lower the extremes and pull the fractions up. This becomes the denominator in a power to raise each element by (sign is retained) where the numerator is 1. This is useful for mellowing out the extremes.

keep.zero

logical. If TRUE the zeros are kept as neutral.

sign

logical. If TRUE the vector will be scaled as (-1, 1, or 0)

...

ignored.

Value

Returns a rescaled vector of the same length as x.

Examples

general_rescale(c(1, 0, -1))
#> [1]  1  0 -1
general_rescale(c(1, 0, -1, 1.4, -2))
#> [1]  0.7145714  0.0000000 -0.5005000  1.0000000 -1.0000000
general_rescale(c(1, 0, -1, 1.4, -2), lower = 0, upper = 1)
#> [1] 0.8823529 0.5882353 0.2941176 1.0000000 0.0000000
general_rescale(c(NA, -4:3))
#> [1]       NA -1.00000 -0.75025 -0.50050 -0.25075  0.00000  0.33400  0.66700
#> [9]  1.00000
general_rescale(c(NA, -4:3), keep.zero = FALSE)
#> [1]         NA -1.0000000 -0.7142857 -0.4285714 -0.1428571  0.1428571  0.4285714
#> [8]  0.7142857  1.0000000
general_rescale(c(NA, -4:3), keep.zero = FALSE, lower = 0, upper = 100)
#> [1]        NA   0.00000  14.28571  28.57143  42.85714  57.14286  71.42857
#> [8]  85.71429 100.00000

## mute extreme values
set.seed(10)
x <- sort(c(NA, -100, -10, 0, rnorm(10, 0, .1), 10, 100), na.last = FALSE)
general_rescale(x)
#>  [1]           NA -1.000000000 -0.100900000 -0.002625046 -0.002369959
#>  [6] -0.002206868 -0.001598569 -0.001363312 -0.001256222 -0.001184068
#> [11]  0.000000000  0.001018727  0.001294251  0.001389405  0.100900000
#> [16]  1.000000000
general_rescale(x, mute = 5)
#>  [1]         NA -1.0000000 -0.6313264 -0.2775830 -0.2682969 -0.2616060
#>  [7] -0.2275036 -0.2059786 -0.1921506 -0.1799156  0.0000000  0.1142786
#> [13]  0.1975151  0.2088417  0.6313264  1.0000000
general_rescale(x, mute = 10)
#>  [1]         NA -1.0000000 -0.7945339 -0.5266486 -0.5177491 -0.5112405
#>  [7] -0.4766860 -0.4535191 -0.4379891 -0.4237726  0.0000000  0.3374006
#> [13]  0.4440785  0.4566685  0.7945339  1.0000000
general_rescale(x, mute = 100)
#>  [1]         NA -1.0000000 -0.9772600 -0.9378682 -0.9362698 -0.9350851
#>  [7] -0.9285578 -0.9239382 -0.9207208 -0.9176839  0.0000000  0.8969724
#> [13]  0.9219944  0.9245786  0.9772600  1.0000000