Get Available Data

See available sentimentr data a data.frame. Note that sentimentr_data is the main function to be used but available_data is exposed to allow other packages to use the functionality in a generic way.

Usage

available_data(regex = NULL, package = "sentimentr", ...)

sentimentr_data(regex = NULL, package = "sentimentr", ...)

Arguments

regex

A regex to search for within the data columns.

package

The name of the package to extract data from.

...

Other arguments passed to grep.

Value

Returns a data.frame

Examples

sentimentr_data()
#>                             Data
#> 1             course_evaluations
#> 2        crowdflower_deflategate
#> 3           crowdflower_products
#> 4  crowdflower_self_driving_cars
#> 5            crowdflower_weather
#> 6                  hotel_reviews
#> 7            hu_liu_apex_reviews
#> 8          hu_liu_cannon_reviews
#> 9         hu_liu_jukebox_reviews
#> 10          hu_liu_nikon_reviews
#> 11          hu_liu_nokia_reviews
#> 12          kaggle_movie_reviews
#> 13  kotzias_reviews_amazon_cells
#> 14          kotzias_reviews_imdb
#> 15          kotzias_reviews_yelp
#> 16                  nyt_articles
#> 17     presidential_debates_2012
#> 18                      sam_i_am
#>                                                              Description
#> 1                                     Student Course Evaluation Comments
#> 2                                   Twitter Tweets About the Deflategate
#> 3                                      Twitter Tweets About the Products
#> 4                                 Twitter Tweets About Self Driving Cars
#> 5                                       Twitter Tweets About the Weather
#> 6                                                          Hotel Reviews
#> 7    Apex AD2600 Progressive-scan DVD player Product Reviews From Amazon
#> 8                           Cannon G3 Camera Product Reviews From Amazon
#> 9  Creative Labs Nomad Jukebox Zen Xtra 40GB Product Reviews From Amazon
#> 10                        Nikon Coolpix 4300 Product Reviews From Amazon
#> 11                                Nokia 6610 Product Reviews From Amazon
#> 12                                                         Movie Reviews
#> 13                                         Kotzias Reviews: Amazon Cells
#> 14                                                 Kotzias Reviews: IMBD
#> 15                                                 Kotzias Reviews: Yelp
#> 16                              Sentiment Scored New York Times Articles
#> 17                                        2012 U.S. Presidential Debates
#> 18                                                         Sam I Am Text
available_data() ## generic version for export
#>                                   Data
#> 1                              cliches
#> 2                         common_names
#> 3       constraining_loughran_mcdonald
#> 4                     emojis_sentiment
#> 5                     freq_first_names
#> 6                      freq_last_names
#> 7                       function_words
#> 8                      grady_augmented
#> 9                          hash_emojis
#> 10              hash_emojis_identifier
#> 11                      hash_emoticons
#> 12                      hash_grady_pos
#> 13                 hash_internet_slang
#> 14                         hash_lemmas
#> 15                   hash_nrc_emotions
#> 16               hash_sentiment_emojis
#> 17                hash_sentiment_huliu
#> 18              hash_sentiment_jockers
#> 19       hash_sentiment_jockers_rinker
#> 20    hash_sentiment_loughran_mcdonald
#> 21                  hash_sentiment_nrc
#> 22            hash_sentiment_senticnet
#> 23            hash_sentiment_sentiword
#> 24              hash_sentiment_slangsd
#> 25         hash_sentiment_socal_google
#> 26               hash_valence_shifters
#> 27                    key_contractions
#> 28 key_corporate_social_responsibility
#> 29                           key_grade
#> 30                          key_rating
#> 31              key_regressive_imagery
#> 32               key_sentiment_jockers
#> 33             modal_loughran_mcdonald
#> 34                        nrc_emotions
#> 35                     pos_action_verb
#> 36              pos_df_irregular_nouns
#> 37                     pos_df_pronouns
#> 38                   pos_interjections
#> 39                     pos_preposition
#> 40                   profanity_alvarez
#> 41                   profanity_arr_bad
#> 42                    profanity_banned
#> 43                    profanity_racist
#> 44                 profanity_zac_anger
#> 45                            sw_dolch
#> 46                          sw_fry_100
#> 47                         sw_fry_1000
#> 48                          sw_fry_200
#> 49                           sw_fry_25
#> 50                          sw_jockers
#> 51           sw_loughran_mcdonald_long
#> 52          sw_loughran_mcdonald_short
#> 53                           sw_lucene
#> 54                           sw_mallet
#> 55                           sw_python
#>                                                                      Description
#> 1                                                                 Common Cliches
#> 2                                                             First Names (U.S.)
#> 3                                           Loughran-McDonald Constraining Words
#> 4                                                           Emoji Sentiment Data
#> 5                                                      Frequent U.S. First Names
#> 6                                                       Frequent U.S. Last Names
#> 7                                                                 Function Words
#> 8  Augmented List of Grady Ward's English Words and Mark Kantrowitz's Names List
#> 9                                                 Emoji Description Lookup Table
#> 10                                                 Emoji Identifier Lookup Table
#> 11                                                                     Emoticons
#> 12                                             Grady Ward's Moby Parts of Speech
#> 13                             List of Internet Slang and Corresponding Meanings
#> 14                                                            Lemmatization List
#> 15                                                             NRC Emotion Table
#> 16                                         Emoji Sentiment Polarity Lookup Table
#> 17                                                  Hu Liu Polarity Lookup Table
#> 18                                              Jockers Sentiment Polarity Table
#> 19                               Combined Jockers & Rinker Polarity Lookup Table
#> 20                                              Loughran-McDonald Polarity Table
#> 21                                                  NRC Sentiment Polarity Table
#> 22                                            Augmented SenticNet Polarity Table
#> 23                                            Augmented Sentiword Polarity Table
#> 24                                              SlangSD Sentiment Polarity Table
#> 25                                                  SO-CAL Google Polarity Table
#> 26                                                              Valence Shifters
#> 27                                                       Contraction Conversions
#> 28  Nadra Pencle and Irina Mălăescu's Corporate Social Responsibility Dictionary
#> 29                                                               Grades Data Set
#> 30                                                              Ratings Data Set
#> 31                      Colin Martindale's English Regressive Imagery Dictionary
#> 32                                                    Jockers Sentiment Data Set
#> 33                                                  Loughran-McDonald Modal List
#> 34                                                                  NRC Emotions
#> 35                                                              Action Word List
#> 36                                                Irregular Nouns Word Dataframe
#> 37                                                                      Pronouns
#> 38                                                                 Interjections
#> 39                                                             Preposition Words
#> 40                                  Alejandro U. Alvarez's List of Profane Words
#> 41                             Stackoverflow user2592414's List of Profane Words
#> 42                                    bannedwordlist.com's List of Profane Words
#> 43                                           Titus Wormer's List of Racist Words
#> 44                                             Zac Anger's List of Profane Words
#> 45                                        Leveled Dolch List of 220 Common Words
#> 46                                    Fry's 100 Most Commonly Used English Words
#> 47                                   Fry's 1000 Most Commonly Used English Words
#> 48                                    Fry's 200 Most Commonly Used English Words
#> 49                                     Fry's 25 Most Commonly Used English Words
#> 50                        Matthew Jocker's Expanded Topic Modeling Stopword List
#> 51                                          Loughran-McDonald Long Stopword List
#> 52                                         Loughran-McDonald Short Stopword List
#> 53                                                          Lucene Stopword List
#> 54                                                          MALLET Stopword List
#> 55                                                          Python Stopword List
available_data(package = 'datasets')
#>                                   Data
#> 1                              cliches
#> 2                         common_names
#> 3       constraining_loughran_mcdonald
#> 4                     emojis_sentiment
#> 5                     freq_first_names
#> 6                      freq_last_names
#> 7                       function_words
#> 8                      grady_augmented
#> 9                          hash_emojis
#> 10              hash_emojis_identifier
#> 11                      hash_emoticons
#> 12                      hash_grady_pos
#> 13                 hash_internet_slang
#> 14                         hash_lemmas
#> 15                   hash_nrc_emotions
#> 16               hash_sentiment_emojis
#> 17                hash_sentiment_huliu
#> 18              hash_sentiment_jockers
#> 19       hash_sentiment_jockers_rinker
#> 20    hash_sentiment_loughran_mcdonald
#> 21                  hash_sentiment_nrc
#> 22            hash_sentiment_senticnet
#> 23            hash_sentiment_sentiword
#> 24              hash_sentiment_slangsd
#> 25         hash_sentiment_socal_google
#> 26               hash_valence_shifters
#> 27                    key_contractions
#> 28 key_corporate_social_responsibility
#> 29                           key_grade
#> 30                          key_rating
#> 31              key_regressive_imagery
#> 32               key_sentiment_jockers
#> 33             modal_loughran_mcdonald
#> 34                        nrc_emotions
#> 35                     pos_action_verb
#> 36              pos_df_irregular_nouns
#> 37                     pos_df_pronouns
#> 38                   pos_interjections
#> 39                     pos_preposition
#> 40                   profanity_alvarez
#> 41                   profanity_arr_bad
#> 42                    profanity_banned
#> 43                    profanity_racist
#> 44                 profanity_zac_anger
#> 45                            sw_dolch
#> 46                          sw_fry_100
#> 47                         sw_fry_1000
#> 48                          sw_fry_200
#> 49                           sw_fry_25
#> 50                          sw_jockers
#> 51           sw_loughran_mcdonald_long
#> 52          sw_loughran_mcdonald_short
#> 53                           sw_lucene
#> 54                           sw_mallet
#> 55                           sw_python
#>                                                                      Description
#> 1                                                                 Common Cliches
#> 2                                                             First Names (U.S.)
#> 3                                           Loughran-McDonald Constraining Words
#> 4                                                           Emoji Sentiment Data
#> 5                                                      Frequent U.S. First Names
#> 6                                                       Frequent U.S. Last Names
#> 7                                                                 Function Words
#> 8  Augmented List of Grady Ward's English Words and Mark Kantrowitz's Names List
#> 9                                                 Emoji Description Lookup Table
#> 10                                                 Emoji Identifier Lookup Table
#> 11                                                                     Emoticons
#> 12                                             Grady Ward's Moby Parts of Speech
#> 13                             List of Internet Slang and Corresponding Meanings
#> 14                                                            Lemmatization List
#> 15                                                             NRC Emotion Table
#> 16                                         Emoji Sentiment Polarity Lookup Table
#> 17                                                  Hu Liu Polarity Lookup Table
#> 18                                              Jockers Sentiment Polarity Table
#> 19                               Combined Jockers & Rinker Polarity Lookup Table
#> 20                                              Loughran-McDonald Polarity Table
#> 21                                                  NRC Sentiment Polarity Table
#> 22                                            Augmented SenticNet Polarity Table
#> 23                                            Augmented Sentiword Polarity Table
#> 24                                              SlangSD Sentiment Polarity Table
#> 25                                                  SO-CAL Google Polarity Table
#> 26                                                              Valence Shifters
#> 27                                                       Contraction Conversions
#> 28  Nadra Pencle and Irina Mălăescu's Corporate Social Responsibility Dictionary
#> 29                                                               Grades Data Set
#> 30                                                              Ratings Data Set
#> 31                      Colin Martindale's English Regressive Imagery Dictionary
#> 32                                                    Jockers Sentiment Data Set
#> 33                                                  Loughran-McDonald Modal List
#> 34                                                                  NRC Emotions
#> 35                                                              Action Word List
#> 36                                                Irregular Nouns Word Dataframe
#> 37                                                                      Pronouns
#> 38                                                                 Interjections
#> 39                                                             Preposition Words
#> 40                                  Alejandro U. Alvarez's List of Profane Words
#> 41                             Stackoverflow user2592414's List of Profane Words
#> 42                                    bannedwordlist.com's List of Profane Words
#> 43                                           Titus Wormer's List of Racist Words
#> 44                                             Zac Anger's List of Profane Words
#> 45                                        Leveled Dolch List of 220 Common Words
#> 46                                    Fry's 100 Most Commonly Used English Words
#> 47                                   Fry's 1000 Most Commonly Used English Words
#> 48                                    Fry's 200 Most Commonly Used English Words
#> 49                                     Fry's 25 Most Commonly Used English Words
#> 50                        Matthew Jocker's Expanded Topic Modeling Stopword List
#> 51                                          Loughran-McDonald Long Stopword List
#> 52                                         Loughran-McDonald Short Stopword List
#> 53                                                          Lucene Stopword List
#> 54                                                          MALLET Stopword List
#> 55                                                          Python Stopword List
sentimentr_data('^hu')
#>                      Data
#> 7     hu_liu_apex_reviews
#> 8   hu_liu_cannon_reviews
#> 9  hu_liu_jukebox_reviews
#> 10   hu_liu_nikon_reviews
#> 11   hu_liu_nokia_reviews
#>                                                              Description
#> 7    Apex AD2600 Progressive-scan DVD player Product Reviews From Amazon
#> 8                           Cannon G3 Camera Product Reviews From Amazon
#> 9  Creative Labs Nomad Jukebox Zen Xtra 40GB Product Reviews From Amazon
#> 10                        Nikon Coolpix 4300 Product Reviews From Amazon
#> 11                                Nokia 6610 Product Reviews From Amazon
sentimentr_data('^(hu|kot)')
#>                            Data
#> 7           hu_liu_apex_reviews
#> 8         hu_liu_cannon_reviews
#> 9        hu_liu_jukebox_reviews
#> 10         hu_liu_nikon_reviews
#> 11         hu_liu_nokia_reviews
#> 13 kotzias_reviews_amazon_cells
#> 14         kotzias_reviews_imdb
#> 15         kotzias_reviews_yelp
#>                                                              Description
#> 7    Apex AD2600 Progressive-scan DVD player Product Reviews From Amazon
#> 8                           Cannon G3 Camera Product Reviews From Amazon
#> 9  Creative Labs Nomad Jukebox Zen Xtra 40GB Product Reviews From Amazon
#> 10                        Nikon Coolpix 4300 Product Reviews From Amazon
#> 11                                Nokia 6610 Product Reviews From Amazon
#> 13                                         Kotzias Reviews: Amazon Cells
#> 14                                                 Kotzias Reviews: IMBD
#> 15                                                 Kotzias Reviews: Yelp
combine_data(sentimentr_data('^(hu|kot)')[[1]])
#>                     source sentiment
#>    1:  hu_liu_apex_reviews         0
#>    2:  hu_liu_apex_reviews         0
#>    3:  hu_liu_apex_reviews         0
#>    4:  hu_liu_apex_reviews         0
#>    5:  hu_liu_apex_reviews         0
#>   ---                               
#> 7089: kotzias_reviews_yelp        -1
#> 7090: kotzias_reviews_yelp        -1
#> 7091: kotzias_reviews_yelp        -1
#> 7092: kotzias_reviews_yelp        -1
#> 7093: kotzias_reviews_yelp        -1
#>                                                                                                                                                                    text
#>    1:                                                                                                             Repost from january 13, 2004 with a better fit title.
#>    2:                                                                                                      Does your apex dvd player only play dvd audio without video?
#>    3:                                                                                                 Or does it play audio and video but scrolling in black and white?
#>    4: Before you try to return the player or waste hours calling apex tech support, or run the player over with your car, try these simple troubleshooting ideas first.
#>    5:                                                                                                                                                       No picture:
#>   ---                                                                                                                                                                  
#> 7089:                                                                                                I think food should have flavor and texture and both were lacking.
#> 7090:                                                                                                                                          Appetite instantly gone.
#> 7091:                                                                                                                Overall I was not impressed and would not go back.
#> 7092:                                                                       The whole experience was underwhelming, and I think we'll just go to Ninja Sushi next time.
#> 7093:                            Then, as if I hadn't wasted enough of my life there, they poured salt in the wound by drawing out the time it took to bring the check.
#>       review_id
#>    1:         1
#>    2:         1
#>    3:         1
#>    4:         1
#>    5:         1
#>   ---          
#> 7089:        NA
#> 7090:        NA
#> 7091:        NA
#> 7092:        NA
#> 7093:        NA

if (FALSE) {
if (!require("pacman")) install.packages("pacman")
pacman::p_load(sentimentr, tidyverse, magrittr)

sentiment_data <- sentimentr_data('^hu') %>%
    pull(Data) %>%
    combine_data() %>%
    mutate(id = seq_len(n())) %>%
    as_tibble()
    
sentiment_test <- sentiment_data %>%
    select(-sentiment) %>%
    get_sentences() %$%
    sentiment(., by = c('id'))

testing <- sentiment_data %>%
    left_join(sentiment_test, by = 'id') %>%
    as_tibble() %>%
    mutate(
        actual = sign(sentiment),
        predicted = sign(ave_sentiment)
    )

testing %$%
    ftable(predicted, actual)
}